refactor: complete autocomplete canonical consolidation
