+++ docs/architecture/python-compatibility-shims.md
@@ -63,7 +63,7 @@ import them.
 | `settings.py` | Explicit direct re-export shim (D-09) | `easyuse_anima.settings.schema`, `.repository`, and `.service` | #163, #186 D-09 | Existing 0.5.2 module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and settings compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and settings migration/round-trip |
 | `storage.py` | Explicit direct re-export shim (D-08) | `easyuse_anima.infrastructure.filesystem.atomic_json` and `.paths` | #163, #186 D-08 | Existing 0.5.2 supported module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and storage compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and last-known-good/atomic-write parity |
 | `autocomplete_index.py` | Explicit direct re-export shim (D-11a) | `easyuse_anima.autocomplete.index` | #162, #186 D-11a | Existing indexed-search surface; exact seven-name `__all__` and identity fixture | External/legacy imports; `autocomplete_dataset.py` now uses the canonical owner | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and index/ranking/rebuild parity |
-| `autocomplete_dataset.py` | Partial explicit compatibility module after D-11b; prompt classification remains root-owned with its D-13 dependency removed | `easyuse_anima.autocomplete.dataset` and `.search`; final classification owner is now unblocked | #162, #186 D-11 | Existing 0.5.2 dataset/search surface canonicalized in PR #386 with direct identity aliases; D-13 switched classification to canonical ANIMA prompt imports | External/legacy imports and prompt tests; `api.py` uses canonical dataset/search and root classification | Unscheduled; final classification identity surface, N+1 gate, and classification/result/API parity |
+| `autocomplete_dataset.py` | Explicit 15-symbol dataset/search/classification identity shim (D-11) | `easyuse_anima.autocomplete.dataset`, `.search`, and `.classification` | #162, #186 D-11 | Existing 0.5.2 dataset/search surface canonicalized in PR #386; final classification Move completed after D-13 removed its root dependency | External/legacy imports and direct compatibility tests; production API uses canonical owners | Unscheduled; first complete canonical+shim release N not yet recorded, then N+1 gate and classification/result/API parity |
 | `wildcard_engine.py` | Partial compatibility module after D-12f2; snapshot lifecycle/expansion remain root-owned | `easyuse_anima.wildcard.models`, `.sources`, `.snapshot`, `.seed`, `.mode`, `.selector`, then lifecycle/expansion modules | #184, #186 D-12 | Existing 0.5.2 models in PR #387, sources in #388, snapshot materialization in #389, seed control in #390, mode contract in #391, selector import contract in #392, and selector Move in #393 with direct aliases | root entrypoint, `nodes.py`, `api.py`, wildcard/workflow tests | Unscheduled; full D-12 move, final identity surface, N+1 gate, and seed/expansion/workflow parity |
 | `prompt_translation.py` | Explicit direct re-export shim (D-01) | `easyuse_anima.translation.*` | #164, #186 D-01 | Existing 0.5.2 supported module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and translation compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and provider-off/API parity |
 | `anima_prompt/` package | Explicit package and submodule identity shims (D-13); D-14 freeze decision pending | `easyuse_anima.prompt.anima.*` | #184, #186 D-13/D-14 | Existing 0.5.2 surface canonicalized in D-13 with explicit public `__all__`, flat/package import parity, and packed closure | External/legacy imports and direct compatibility tests; production callers use the canonical package | Unscheduled; first canonical+shim release N not yet recorded, then D-14/N+1 gate and prompt correction/parser parity |
