+++ tools/analyze_python_backend.py
@@ -39,6 +39,7 @@ REGISTRY_ENTRY_MODULE_CANDIDATES = (
     "anima_prompt/normalize.py",
     "anima_prompt/ordering.py",
     "anima_prompt/parser.py",
+    "autocomplete_dataset.py",
     "autocomplete_index.py",
     "nodes.py",
     "prompt_translation.py",
